move dialog to fix containment issue
